+++ src/contexts/Farming/Context.ts
@@ -3,8 +3,8 @@ import { createContext } from "react";
 import { ContextValues } from "./types";
 
 const Context = createContext<ContextValues>({
-  farmingStartTime: [1600545500000],
-  setConfirmTxModalIsOpen: () => {},
+  farmingStartTime: [1],
+  setConfirmModal: () => {},
 });
 
 export default Context;